Небольшие переименования в соответствии с именованием тестовых классов
* SerializationTests.cpp -> IntsTest.cpp;
* Для тестов сериализации целых чисел добавлен namespace ints;

+ Исправлена опечатка в названии функии (un)serialize в IntsTest.cpp (ex SerializationTests.cpp);
