+++ main/main.cpp
@@ -3,8 +3,8 @@
 
 int main()
 {
-	serialization_tests::ints();
-	serialization_tests::nested_int_structs();
+	serialization_tests::ints::structs_with_ints();
+	serialization_tests::ints::nested_int_structs();
 
 	return 0;
 }
