+++ tests/SerializationTests.h
@@ -2,7 +2,11 @@
 
 namespace serialization_tests
 {
-	void ints();
-	void nested_int_structs();
+	namespace ints
+	{
+		void structs_with_ints();
+		void nested_int_structs();
+
+	} // namespace ints
 
 } // namespace serialization_tests
